backend/requirements.txt
apparmor==3.1.1
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.2
autocommand==2.2.1
btrfsutil==5.19.1
catfish==4.16.4
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
cryptography==38.0.3
Cython==0.29.32
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
filelock==3.8.0
gufw==22.4.0
idna==3.4
inflect==6.0.0
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
keyutils==0.6
lensfun==0.3.3
LibAppArmor==3.1.1
lightdm-gtk-greeter-settings==1.2.2
lit==14.0.6.dev0
manjaro-sdk==0.1
menulibre==2.2.3
more-itertools==8.13.0
mugshot==0.4.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
oauthlib==3.2.2
ordered-set==4.1.0
packaging==21.3
pacman-mirrors==4.23.2
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
psutil==5.9.2
psycopg2==2.9.3
ptyprocess==0.7.0
pycairo==1.21.0
pycparser==2.21
pydantic==1.10.2
PyGObject==3.42.2
PyJWT==1.7.1
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.2
pysmbc==1.0.23
python3-openid==3.2.0
pytz==2022.6
pyxdg==0.28
PyYAML==6.0
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
shiboken6==6.3.2
six==1.16.0
sqlparse==0.4.3
systemd-python==235
team==1.0
tomli==2.0.1
trove-classifiers==2022.9.26
typing_extensions==4.3.0
udiskie==2.4.2
ufw==0.36.1
urllib3==1.26.12
validate-pyproject==0.10.1
virtualenv==20.16.5
